# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Node Express Compass Uol",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "User",
      "description": "Endpoints"
    },
    {
      "name": "Product",
      "description": "Endpoints"
    },
    {
      "name": "Joke",
      "description": "Endpoints"
    }
  ],
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/health": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/joke": {
      "get": {
        "tags": [
          "Joke"
        ],
        "description": "Endpoint para obter piadas aleatórias.",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/auth": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Auth JWT, username: admin, password: 123456",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/products/": {
      "get": {
        "tags": [
          "Product"
        ],
        "description": "Endpoint para obter todos os produto.",
        "responses": {
          "200": {
            "description": "Produtos encontrados.",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Product"
        ],
        "description": "Endpoint para criar um produto.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "newProduct",
            "in": "body",
            "description": "Informações do produto.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/addProduct"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produto criado.",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/products/{id}": {
      "get": {
        "tags": [
          "Product"
        ],
        "description": "Endpoint para obter um produto.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do produto."
          }
        ],
        "responses": {
          "200": {
            "description": "Produto encontrado.",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Product"
        ],
        "description": "Endpoint para alterar um produto.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do produto."
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "updateProduct",
            "in": "body",
            "description": "Informações do produto.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/addProduct"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produto alterado.",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Product"
        ],
        "description": "Endpoint para apagar um produto.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID do produto."
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Produto apagado.",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Jhon Doe"
        }
      }
    },
    "Product": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Sabonete"
        },
        "description": {
          "type": "string",
          "example": "Caixa de sabonete com 10 unidades"
        },
        "price": {
          "type": "number",
          "example": 20
        },
        "createdAt": {
          "type": "string",
          "example": "2022-12-22T12:18:06.844+0530"
        },
        "updatedAt": {
          "type": "string",
          "example": "2022-12-22T12:18:06.844+0530"
        },
        "createdBy": {
          "type": "number",
          "example": 1
        },
        "updatedBy": {
          "type": "number",
          "example": 1
        }
      },
      "required": [
        "name",
        "price"
      ]
    },
    "addProduct": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Sabonete"
        },
        "description": {
          "type": "string",
          "example": "Caixa de sabonete com 10 unidades"
        },
        "price": {
          "type": "number",
          "example": 20
        }
      },
      "required": [
        "name",
        "price"
      ]
    }
  }
}